fck-nat-fastapi/requirements.txt
aws-cdk-lib==2.135.0
constructs>=10.0.0,<11.0.0
cdk-fck-nat==1.3.15
